Fixed Warnings

+++ LectureConnect/SocketRocket/Socket.io/SocketIO.m
@@ -590,7 +590,7 @@ NSString* const SocketIOException = @"SocketIOException";
     // check for server status code (http://gigliwood.com/weblog/Cocoa/Q__When_is_an_conne.html)
     if ([response respondsToSelector:@selector(statusCode)]) {
         NSInteger statusCode = [((NSHTTPURLResponse *)response) statusCode];
-        DEBUGLOG(@"didReceiveResponse() %i", statusCode);
+        DEBUGLOG(@"didReceiveResponse() %li", (long)statusCode);
         
         if (statusCode >= 400) {
             // stop connecting; no more delegate messages
